+++ src/main/java/com/calibermc/caliber/block/management/CaliberBlockManager.java
@@ -2,6 +2,7 @@ package com.calibermc.caliber.block.management;
 
 import com.calibermc.caliber.block.CaliberBlocks;
 import com.calibermc.caliber.compat.block.BiomesOPlentyBlocks;
+import com.calibermc.caliber.compat.block.CreateBlocks;
 import com.calibermc.caliber.compat.block.RegionsUnexploredBlocks;
 import com.calibermc.caliberlib.block.management.ModBlockHelper;
 import com.calibermc.caliberlib.block.properties.BlockProps;
@@ -52,6 +53,17 @@ public class CaliberBlockManager {
 
             });
         }, BiomesOPlentyBlocks.BLOCKS, properties, blockSupplier, variants).registerBlockFunc(BiomesOPlentyBlocks::registerBlock).build();
+    }
+
+    public static BlockManager registerCreate(String name, Supplier<BlockBehaviour.Properties> properties, Supplier<Block> blockSupplier, Collection<ModBlockFamily.Variant> variants) {
+        return BlockManager.register(name, (b) -> {
+            BlockSetType.values().forEach((type) -> {
+                if (name.contains(type.name())) {
+                    b.type(type);
+                }
+
+            });
+        }, CreateBlocks.BLOCKS, properties, blockSupplier, variants).registerBlockFunc(CreateBlocks::registerBlock).build();
     }
 
     public static BlockManager registerRU(String name, Supplier<BlockBehaviour.Properties> properties, Supplier<Block> blockSupplier, Collection<ModBlockFamily.Variant> variants) {