+++ src/main/java/com/calibermc/caliber/compat/item/CreateItems.java
@@ -1,0 +1,23 @@
+package com.calibermc.caliber.compat.item;
+
+import net.minecraft.world.item.Item;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+
+public class CreateItems {
+    public static int itemCount = 0;
+
+    public static final DeferredRegister<Item> ITEMS =
+            DeferredRegister.create(ForgeRegistries.ITEMS, "create");
+
+
+    public static void printItemCounts() {
+        System.out.println("Create Items Registered: " + itemCount);
+    }
+    public static void register(IEventBus eventBus) {
+        itemCount ++;
+        printItemCounts();
+        ITEMS.register(eventBus);
+    }
+}