+++ src/main/java/com/calibermc/caliber/data/CaliberBlockFamilies.java
@@ -665,7 +665,7 @@ public class CaliberBlockFamilies {
     // COMPAT
     static {
         if (ModCompats.BOP) MAP.putAll(BiomesOPlentyBlockFamilies.getAllFamiliesAsMap());
-//        if (ModCompats.CREATE) MAP.putAll(CreateBlockFamilies.getAllFamiliesAsMap());
+        if (ModCompats.CREATE) MAP.putAll(CreateBlockFamilies.getAllFamiliesAsMap());
         if (ModCompats.REGIONS_UNEXPLORED) MAP.putAll(RegionsUnexploredBlockFamilies.getAllFamiliesAsMap());
 
     }