+++ src/main/java/com/calibermc/caliber/item/CreativeTabs.java
@@ -3,6 +3,7 @@ package com.calibermc.caliber.item;
 import com.calibermc.caliber.Caliber;
 import com.calibermc.caliber.block.CaliberBlocks;
 import com.calibermc.caliber.compat.item.BiomesOPlentyItems;
+import com.calibermc.caliber.compat.item.CreateItems;
 import com.calibermc.caliber.compat.item.RegionsUnexploredItems;
 import com.calibermc.caliber.compat.ModCompats;
 import net.minecraft.core.registries.Registries;
@@ -27,6 +28,11 @@ public class CreativeTabs {
                         }
                         if (ModCompats.BOP) {
                             for (RegistryObject<Item> entry : BiomesOPlentyItems.ITEMS.getEntries()) {
+                                pOutput.accept(entry.get());
+                            }
+                        }
+                        if (ModCompats.CREATE) {
+                            for (RegistryObject<Item> entry : CreateItems.ITEMS.getEntries()) {
                                 pOutput.accept(entry.get());
                             }
                         }