+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ### 0.2.4 (in progress)
+#### Features
 - Added vanilla style slabs for all variants that don't have them
 - Added Chiseled Polished Blackstone variants
 - Added Cracked Polished Blackstone variants
@@ -12,20 +13,27 @@
 - Added Cracked Stone Brick variants
 - Added Chiseled Sandstone variants
 - Added Chiseled Red Sandstone variants
+- Added Purpur Pillar Variants
 - Added Regions Unexplored Chalk Pillar variants
 - Renamed Regions Unexplored Chalk variants to Chalk Block _(this is to allow the Chalk Pillar variants to be added)_
-- Added mod compatibility variants for Create
-- Added mod compatibility variants for Stone Works
-- Added mod compatibility variants for Quark
-- Added mod compatibility variants for Ice and Fire
-- Bugfix: removed duplicate item models for mod compat variants _(this was causing some items to appear invisible in the inventory)_
+- Added BOP Smooth (Black, Orange, White) Sandstone variants
+- Added mod compatibility variants for Create Natural Stone types
 
+[//]: # (- Added mod compatibility variants for Stone Works)
+[//]: # (- Added mod compatibility variants for Quark)
+[//]: # (- Added mod compatibility variants for Ice and Fire)
+
+#### Bugfixes
+- Removed duplicate item models for mod compat variants _(this was causing some items to appear invisible in the inventory)_
+
+---
 ### 0.2.3
 - Downgraded Minimum Forge Version <- 47.1.3 to allow NeoForge 1.20.1 support
 - Added missing recipes for Diagonal Beams
 - Added Mossy Cobblestone variants
 - Added Mossy Stone Brick variants
 - Added Smooth Stone variants
+- Added Wool and Terracotta Slabs
 - Added mod compatibility variants for Biomes O' Plenty
 - Added mod compatibility variants for Regions Unexplored
 - Added missing translations for en_us
@@ -34,9 +42,11 @@
 - Bugfix: fixed Polished Granite BlockPicker displaying Granite stairs and walls, (now displays Polished Granite variants)
 - Bugfix: fixed ArchBlock textures not displaying correctly in all situations (there may still be some glitches in rare case)
 
+---
 ### 0.2.2
 - Added Diagonal Beam variants for all Stripped Wood
 
+---
 ### 0.2.1
 - Slab variants are now just called Layers (Sorry if this breaks any builds, but it just makes sense to avoid confusion with vanilla slabs)
 - Vertical Slab variants are now just called Vertical Layers
@@ -53,6 +63,7 @@
 - Removed dulicate wooden Pressure Plate variants
 - Added Tags data gen to ensure all Minecraft, Forge and Custom tags are accounted for
 
+---
 ### 0.2.0
 - Added Forge 1.20.1 support
 - Removed Sorting Buttons (are included in Buildify now)
@@ -65,12 +76,14 @@
 - Bugfix: fixed red_sand_layer texture showing as sand
 - Bugfix: fixed slab_layer removing block from inventory on placement of double_slab in Slab Mode
 
+---
 ### 0.1.3
 - Updated Buildify Dependency Version to 0.1.3+
 - Adjusted placement of Sorting Buttons when using Chest
 - Bugfix: fixed issue that was preventing startup of dedicated servers
 - Bugfix: fixed villagers not able to use Tall Doors
 
+---
 ### 0.1.2
 - Added a total of 14 new blocks!
 - Current Mod Block Count is at 5,820
@@ -87,6 +100,7 @@
 - Bugfix: fixed hitbox appearing in wrong place for Quarter Blocks with a single layer facing
 - Bugfix: adjusted hitboxes for Layered Blocks to Reduce Ambient Occlusion issues
 
+---
 ### 0.1.1
 - Added a total of 29 new blocks!
 - Current Mod Block Count is at 5,806
@@ -99,11 +113,13 @@
 - Bugfix: fixed lighting/shading on 67 degree roof blocks
 - Bugfix: fixed Sandstone, Red Sandstone, Smooth Sandstone and Quartz Block Pickers showing Black Marble Variants
 
+---
 ### 0.1.0
 - BETA RELEASE
 - While this is a beta release, it is still very much a work in progress.  The mod is stable with no known bugs at this point. There are still many blocks to add and some bugs may still exist.
 - Added finishing touches on the recipe generation system (this will ensure that now recipes are missed in future updates)
 
+---
 ### 0.0.31
 - Added a total of 1,509 new blocks!
 - Current Mod Block Count is at 5,777 !!!!!! (this is a huge update)
@@ -136,6 +152,7 @@
 - Major Bugfix: fixed all block textures (blocks that had different top, bottom, sides were not generating correctly)
 - BETA RELEASE IS COMING SOON!
 
+---
 ### 0.0.30
 - Added a total of 645 new blocks!
 - Current Mod Block Count is at 4,268
@@ -161,10 +178,12 @@
 - Bugfix: fixed visual glitch with Arrowslit Blocks
 - Bugfix: fixed occulsion/shadows on 67 degree roof blocks (this is experimental and may be reworked in the future)
 
+---
 ### 0.0.29
 - BugFix: fixed grass_stems not dropping from grass
 - Removed Buildify Features from Caliber-Mod to add greater mod compatibility for Block Picker feature, added as a dependency/library for Caliber-Mod
 
+---
 ### 0.0.28
 - Added a total of 42 new blocks!
 - Current Mod Block Count is at 3,623
@@ -175,6 +194,7 @@
 - Added Stained Planks Recipes
 - Added Roofing Recipes
 
+---
 ### 0.0.27
 - Added a total of 214 new blocks!
 - Current Mod Block Count is at 3,581
@@ -187,6 +207,7 @@
 - Added all current Kiln recipes
 - All missing recipes will be added asap
 
+---
 ### 0.0.26
 - Major Code Optimization (Removed over 10k lines of code and prevented another 30k+ by implementing a new system for registering blocks and generating data/assets.
   This should allow for faster updates and easier maintenance of the mod going forward)
@@ -202,9 +223,11 @@
 - Added all variants of 67 degree roofs
 - Added all variants of roof peaks
 
+---
 ### 0.0.25
 - Bugfix: fixed major bug preventing startup of dedicated servers on last update 0.0.24
 
+---
 ### 0.0.24
 - Added a total of 261 new blocks!
 - Current Mod Block Count is at 3,233
@@ -228,6 +251,7 @@
 - Bugfix: adjusted inventory logic for Slab Buildmode.  When placing slab thickness blocks the correct inventory amounts will be consumed.
 - Bugfix: fixed double eighth blocks not placing correctly
 
+---
 ### 0.0.23
 - Updated Block Picker GUI textures
 - Updated Blackstone textures
@@ -236,6 +260,7 @@
 - Bugfix: fixed QuarterLayerBlock hitboxes and textures
 - Bugfix: fixed Polished Blackstone Bricks and Polished Blackstone QuarterLayerBlock models (top models were placing on bottom)
 
+---
 ### 0.0.22
 - Added a total of 341 new blocks!
 - Current Mod Block Count is at 2,972
@@ -246,6 +271,7 @@
 - Added all variants of Quartz and Smooth Quartz
 - Added all variants of Smooth Dark Limestone, Smooth Light Limestone, Smooth Limestone, Smooth Pink Limesotne & Smooth Tan Limestone
 
+---
 ### 0.0.21
 - Improved performance of slab/layer buildmodes
 - Reworked ArchBlock and LargeArchBlock blockstates, models, textures, hitboxes
@@ -255,6 +281,7 @@
 - Bugfix: fixed crash when using Hammer Item on layered blocks
 - Bugfix: fixed "out of nails" message when using Hammer in Creative Mode.
 
+---
 ### 0.0.20
 - Added sounds for Hammer Item
 - Added buildmode command to change behaviour of all Layered Blocks (caliber.buildmode)
@@ -265,6 +292,7 @@
   - Creative: left click to change property and right click to cycle states
   - Survival: sneak+right click to change blockstate property and right click to cycle states (1 nail per interaction will be consumed)
 
+---
 ### 0.0.19
 - Added a total of 1,067 new blocks!
 - Current Mod Block Count is at 2,631
@@ -288,6 +316,7 @@
 - Removed unused texture files
 - Updated Silver and Tin Textures
 
+---
 ### 0.0.18
 - Current Mod Block Count is at 1,564
 - Added new trimmed BlockState for Arch Blocks to prevent most texture issues
@@ -313,29 +342,35 @@
 - Bugfix: fixed Corner Slab hitboxes
 - Bugfix: fixed pressure plate tags
 
+---
 ### 0.0.17
 - Added all Mossy Cobbled Limestone variants (were missing from last update)
 - Added Mod Compat to all Creative Tabs via Tags (modders can now add their blocks to CaliberMod Tabs)
 - Lots of code cleanup
 - Bugfix: adjusted particle textures for all Arch Block variants
 
+---
 ### 0.0.16
 - Added all existing block variants of HalfWindow
 - Added all Mossy Stone Brick variants
 - Added all Mossy Cobbled Limestone variants
 
+---
 ### 0.0.15
 - Added all block variants of Cut Sandstone and Smooth Sandstone
 - Block Picker overhaul allowing for shift+click, right click and dropping blocks onto GUI to remove
 
+---
 ### 0.0.14
 - Bugfix: Copy Block key now only works in Creative Mode
 
+---
 ### 0.0.13
 - Added HalfWindow Shape
 - Added all shape variants for all vanilla wood planks
 - Bugfix: waterlogged blocks now function properly
 
+---
 ### 0.0.12
 - Added tags for all existing blocks
 - Bugfix: fixed server sync issue with Block Picker
@@ -343,6 +378,7 @@
 - Bugfix: added all missing model files, corrected blockstates, fixed textures
 - Bugfix: added new enums to fix missing blockstate errors in console
 
+---
 ### 0.0.11
 - Added Block Picker GUI
 - Added Custom Creative Inventory
@@ -350,6 +386,7 @@
 - dded Block/BlockState Copy Key
 - Bugfix: Added missing Limestone Wall
 
+---
 ### 0.0.10
 - Added Arch Blocks!
 - Added all Natural Stone variants of Arch Block
@@ -361,21 +398,26 @@
 - Bug Fix: Adjusted ArrowSlit Blocks hitboxes and outlines
 - Cleaned up a bunch of code!
 
+---
 ### 0.0.9
 - Added all Natural Stone variants of CornerSlab and VerticalCornerSlab.
 
+---
 ### 0.0.8
 - Added all Natural Stone Variants of Arrow Slit Block, Balustrade Block, Capital Block and Window Block -Added All current custom blocks to Red Sandstone & Sandstone
 - Fixed Textures on Custom Sandstone blocks
 - Added translations and loot for all current blocks
 
+---
 ### 0.0.7
 - Added Corner Block
 - Added all except cobble and sandstone Corner Blocks (will be included in the next update)
 
+---
 ### 0.0.6
 - Bug Fix: Adjusted rotation of Vertical Quarters
 
+---
 ### 0.0.5
 - Added World Gen for Natural Stones
 - Added all variations of Natural Stone Vertical Slab, Quarter, Vertical Quarter and Eighth Blocks.
@@ -383,12 +425,15 @@
 - Lots of code optimizations
 - Bug Fixes
 
+---
 ### 0.0.4
 - not released
 
+---
 ### 0.0.3
 - not released
 
+---
 ### 0.0.2
 - Initial Alpha Release
 - This mod is in the early development phase